+++ src/components/com_settings/dispatcher.php
@@ -5,9 +5,11 @@ class ComSettingsDispatcher extends ComBaseDispatcherDefault
     protected function _actionDispatch(AnCommandContext $context)
     {
         if ($this->getController()->getView()->getName() == 'settings') {
-            $this->getController()->execute('read', $context);
+            $this->setController('config')
+            ->getController()
+            ->setView('config');
         }
-
+        
         return parent::_actionDispatch($context);
     }
 }