+++ src/libraries/anahita/log/log.php
@@ -66,7 +66,7 @@ class AnLog extends AnObject implements AnServiceInstantiatable
     */
     protected function _initialize(AnConfig $config)
     {
-         $settings = $this->getService('com:settings.setting');
+         $settings = $this->getService('com:settings.config');
 
          $config->append(array(
             'file' => 'error.php',